// devui/src/app/modules/dnb/models/path/api-path.constant.ts
export enum apiMap {
  restServiceDnbAggregator = "dnb-aggregator-api/",
  restServiceDnbAuth = "dnb-auth-api/",
  restServiceDnbAPI = "dnb-drd-api/",
}

export enum apiPath {
  document = "document/",
  documents = "documents/",
  section = "/section/",
  sections = "/sections/",
  users = "users/",
  drugVersionLatest = "drugs/versions/latest/",
  drugVersionList = "drugs/versions/list/",
  drugVersionInApproval = "drugs/versions/approval-process",
  myWork = "my-work",
  drugListByPage = "drugs/list-by-page",
  authUsers = "auth/users/",
  authToken = "auth/token",
  drugFilter = "drugs/filter-drugs",
  download = "documents/",
  drugVersion = "drug-version/",
  addDrugVersion = "drugs/versions",
  roles = "roles/",
  permissions = "/permissions",
  drugs = "drugs",
  midRulesTemplates = "mid-rules/templates",
  midRulesSearch = "mid-rules/search",
  midRulesAdmin = "mid-rules/templates/all",
  submit = "_submit",
  save = "_save",
  approve = "_approve",
  approvalProcess = "approval-process/",
  submitReview = "/_submit-for-review",
  tabInProgress = "in-progress",
  submittedForReview = "submitted-for-review",
  feedBack = "/_feedback",
  feedBackFlow = "feedback",
  verify = "/verify/",
  reclaim = "/_reclaim",
  approverClaim = "drugs/versions/approver-claim/",
  inReview = "in-review",
  DNBE = "DNBE/",
  DNBA = "DNBA/",
  DNBEA = "DNBEA/",
  DNBAdminToEditor = "DNBADMIN-to-DNBE/",
  reassign = "_reassign",
  DNBAdminToApprover = "DNBADMIN-to-DNBA/",
  DNBADMIN = "DNBADMIN/",
  DNBEDNBA = "DNBE&DNBA/",
  Icd10CodeAutopopulate = "icd10code/list",
  dnbTemplate = "dbn-template",
  dnbGenerateTemplate = "generate-dnb-edit-file",
  manualIngestion = "manual-ingestion/",
  downloadIngest = "download",
  ingest = "ingest",
  getMidRuleDiffTemplate = "diff-input-template",
  generateReportByFile = "read-diff-input-template",
  generateReportByRules = "midrule-diff-report",
  getELLRules = "ell-policies/list/",
  validateMidRules = "validate-midrule-version",
  pinnedSections = "/dnb/pinned-sections",
  ellMappings = "mappings/ell",
  ellRulesTopicName = "drug/topic/?topicName=",
  ellRulesDrugCode = "drug/topic/?drugCode=",
}